Add multi-part secret handling to aws_secrets.py
If the generated Gluu secrets (their compressed binary value) would be too large to fit into one secret, new secrets would be created to hold the rest of the data. At the same time a simple header is inserted to indicate the number of secrets used to store the binary. The secrets' name are created with a postfix indicating the part number (gluu_secrets_1 to gluu_secrets_n).